// types/Meeting.ts
export type Meeting = {
  id_bigint: string;
  worldid_mixed: string;
  service_body_bigint: string;
  weekday_tinyint: string;
  venue_type: string;
  start_time: string;
  duration_time: string;
  time_zone: string;
  formats: string;
  longitude: string;
  latitude: string;
  root_server_uri: string;
  format_shared_id_list: string;
  meeting_name: string;
  location_text: string;
  location_info: string | undefined;
  location_street: string;
  location_neighborhood: string;
  location_municipality: string;
  location_sub_province: string;
  location_province: string;
  location_postal_code_1: string;
  comments: string;
  contact_phone_2: string;
  contact_email_2: string;
  contact_name_2: string;
  contact_phone_1: string;
  contact_email_1: string;
  contact_name_1: string;
  phone_meeting_number: string;
  virtual_meeting_link: string;
  virtual_meeting_additional_info: string;
};